requirements.txt
requests
python-docx
PyPDF2
reportlab
